First split of value-trait
