+++ src/known_key.rs
@@ -1,4 +1,4 @@
-use crate::{MutableValue, Value as ValueTrait, ValueType};
+use crate::{Mutable, Value as ValueTrait, ValueType};
 use halfbrown::RawEntryMut;
 use std::borrow::{Borrow, Cow};
 use std::fmt;
@@ -102,8 +102,8 @@ impl<'key> KnownKey<'key> {
     where
         'key: 'value,
         'value: 'borrow,
-        V: MutableValue + 'value,
-        <V as MutableValue>::Key: Hash + Eq + Borrow<str>,
+        V: Mutable + 'value,
+        <V as Mutable>::Key: Hash + Eq + Borrow<str>,
     {
         target.as_object_mut().and_then(|m| {
             match m
@@ -153,8 +153,8 @@ impl<'key> KnownKey<'key> {
     where
         'key: 'value,
         'value: 'borrow,
-        V: ValueTrait + MutableValue + 'value,
-        <V as MutableValue>::Key: Hash + Eq + Borrow<str> + From<Cow<'key, str>>,
+        V: ValueTrait + Mutable + 'value,
+        <V as Mutable>::Key: Hash + Eq + Borrow<str> + From<Cow<'key, str>>,
         F: FnOnce() -> V,
     {
         if !target.is_object() {
@@ -204,8 +204,8 @@ impl<'key> KnownKey<'key> {
     where
         'key: 'value,
         'value: 'borrow,
-        V: MutableValue + 'value,
-        <V as MutableValue>::Key: Hash + Eq + Borrow<str> + From<Cow<'key, str>>,
+        V: Mutable + 'value,
+        <V as Mutable>::Key: Hash + Eq + Borrow<str> + From<Cow<'key, str>>,
     {
         if !target.is_object() {
             return Err(Error::NotAnObject(target.value_type()));
@@ -231,7 +231,7 @@ mod tests {
     #![allow(clippy::unnecessary_operation, clippy::non_ascii_literal)]
     use super::*;
     use crate::borrowed::*;
-    use crate::{BorrowedValue, Value as ValueTrait, ValueBuilder};
+    use crate::{BorrowedValue, Builder, Value as ValueTrait};
 
     #[test]
     fn known_key() {
