+++ src/lib.rs
@@ -217,6 +217,7 @@ pub mod value;
 
 use std::mem;
 use std::str;
+pub use value_trait::StaticNode;
 
 pub use crate::error::{Error, ErrorType};
 pub use crate::value::*;
@@ -229,7 +230,7 @@ mod known_key;
 #[cfg(feature = "known-key")]
 pub use known_key::{Error as KnownKeyError, KnownKey};
 
-pub use crate::tape::{Node, StaticNode, Tape};
+pub use crate::tape::{Node, Tape};
 
 /// Creates a tape from the input for later consumption
 /// # Errors
@@ -641,10 +642,10 @@ impl<'de> Deserializer<'de> {
 #[cfg(test)]
 mod tests {
     #![allow(clippy::unnecessary_operation, clippy::non_ascii_literal)]
-    use super::WritableValue;
     use super::{owned::Value, to_borrowed_value, to_owned_value, Deserializer};
     use crate::tape::*;
     use proptest::prelude::*;
+    use value_trait::{StaticNode, Writable};
 
     #[test]
     fn count1() {
@@ -692,7 +693,7 @@ mod tests {
     #[test]
     fn odd_nuber() {
         use super::value::owned::to_value;
-        use super::value::{MutableValue, ValueBuilder};
+        use super::value::{Builder, Mutable};
 
         let mut d = String::from(
             r#"{"name": "max_unsafe_auto_id_timestamp", "value": -9223372036854776000}"#,
@@ -711,7 +712,7 @@ mod tests {
     #[test]
     fn odd_nuber2() {
         use super::value::owned::to_value;
-        use super::value::{MutableValue, ValueBuilder};
+        use super::value::{Builder, Mutable};
 
         let mut d = String::from(
             r#"{"name": "max_unsafe_auto_id_timestamp", "value": 9223372036854776000}"#,
@@ -840,13 +841,12 @@ mod tests {
 mod tests_serde {
     #![allow(clippy::unnecessary_operation, clippy::non_ascii_literal)]
     use super::serde::from_slice;
-    use super::value::{MutableValue, ValueBuilder};
     use super::{owned::to_value, owned::Object, owned::Value, to_borrowed_value, to_owned_value};
-    use crate::tape::*;
     use halfbrown::HashMap;
     use proptest::prelude::*;
     use serde::Deserialize;
     use serde_json;
+    use value_trait::{Builder, Mutable, StaticNode};
 
     #[test]
     fn empty() {
