+++ src/prelude.rs
@@ -1,2 +1,2 @@
 /// Prelude to include needed traits
-pub use super::{MutableValue, Value as ValueTrait, ValueBuilder};
+pub use value_trait::{AccessError, Builder, Mutable, Value as ValueTrait, ValueType, Writable};
