+++ src/sse42/mod.rs
@@ -1,4 +1,3 @@
 pub mod deser;
-pub mod generator;
 pub mod stage1;
 pub mod utf8check;
