+++ src/value/owned/serialize.rs
@@ -5,15 +5,15 @@
 // https://github.com/maciejhirsz/json-rust/blob/master/src/codegen.rs
 
 use super::{Object, Value};
-use crate::value::generator::*;
-use crate::value::{Value as ValueTrait, WritableValue};
+use crate::prelude::*;
 use crate::{stry, StaticNode};
 use std::io;
 use std::io::Write;
+use value_trait::generator::*;
 
 //use util::print_dec;
 
-impl WritableValue for Value {
+impl Writable for Value {
     #[inline]
     fn encode(&self) -> String {
         let mut g = DumpGenerator::new();
@@ -151,8 +151,8 @@ where
 #[cfg(test)]
 mod test {
     use super::Value;
+    use crate::prelude::*;
     use crate::StaticNode;
-    use crate::WritableValue;
     #[test]
     fn null() {
         assert_eq!(Value::Static(StaticNode::Null).encode(), "null")
