+++ value-trait/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "value-trait"
+version = "0.1.0"
+authors = ["Heinz N. Gies <heinz@licenser.net>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+itoa = "0.4"
+ryu = "1"
+halfbrown = "0.1"
+float-cmp = "0.6"
